+++ packages/stream_chat_flutter/CHANGELOG.md
@@ -1,4 +1,9 @@
-## Upcoming
+## 4.1.0
+
+✅ Added
+
+- [[#1119]](https://github.com/GetStream/stream-chat-flutter/issues/1119) Added an option to disable mentions overlay in `StreamMessageInput`
+- Deprecated `disableEmojiSuggestionsOverlay` in favor of `enableEmojiSuggestionsOverlay` in `StreamMessageInput`
 
 🐞 Fixed
 
@@ -6,6 +11,10 @@
 - Fixed StreamChannelHeader and StreamThreadHeader subtitle alignment.
 - Fixed message widget thread indicator in reverse mode.
 - [[#1044]](https://github.com/GetStream/stream-chat-flutter/issues/1044): Refactor StreamMessageWidget bottom row to use Text.rich.
+
+🔄 Changed
+
+- Removed `isOwner` condition from `ChannelBottomSheet` and `StreamChannelInfoBottomSheet` for delete option tile.
 
 ## 4.0.1